Face Reconstruction/3D Face Reconstruction with Weakly-Supervised Learning/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
blas=1.0=mkl
ca-certificates=2020.11.8=h5b45459_0
certifi=2020.11.8=py38haa244fe_0
chardet=3.0.4=pypi_0
cudatoolkit=10.1.243=h74a9793_0
docopt=0.6.2=pypi_0
freetype=2.10.4=hd328e21_0
future=0.18.2=pypi_0
fvcore=0.1.2.post20201111=pypi_0
icc_rt=2020.2=intel_254
idna=2.10=pypi_0
intel-openmp=2020.2=254
jpeg=9b=hb83a4c4_2
libpng=1.6.37=h2a8f88b_0
libtiff=4.1.0=h56a325e_1
lz4-c=1.9.2=hf4a77e7_3
mkl=2020.2=256
mkl-service=2.3.0=py38hb782905_0
mkl_fft=1.2.0=py38h45dec08_0
mkl_random=1.1.1=py38h47e9c7a_0
msys2-conda-epoch=20160418=1
ninja=1.7.2=0
numpy=1.19.2=py38hadc3359_0
numpy-base=1.19.2=py38ha3acd2a_0
olefile=0.46=py_0
openssl=1.1.1h=he774522_0
pillow=8.0.1=py38h4fa10fc_0
pip=20.2.4=py38haa95532_0
pipreqs=0.4.10=pypi_0
portalocker=2.0.0=pypi_0
python=3.8.5=h5fd99cc_1
python_abi=3.8=1_cp38
pytorch=1.6.0=py3.8_cuda101_cudnn7_0
pytorch3d=0.3.0=pypi_0
pywin32=300=pypi_0
pyyaml=5.3.1=pypi_0
requests=2.25.0=pypi_0
scipy=1.5.2=py38h14eb087_0
setuptools=50.3.1=py38haa95532_1
six=1.15.0=py38haa95532_0
sqlite=3.33.0=h2a8f88b_0
tabulate=0.8.7=pyh9f0ad1d_0
termcolor=1.1.0=pypi_0
tk=8.6.10=he774522_0
torchvision=0.7.0=py38_cu101
tqdm=4.52.0=pyhd3deb0d_0
urllib3=1.26.2=pypi_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_3
wheel=0.35.1=pyhd3eb1b0_0
wincertstore=0.2=py38_0
xz=5.2.5=h62dcd97_0
yacs=0.1.8=pypi_0
yaml=0.2.5=he774522_0
yarg=0.1.9=pypi_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.5=h04227a9_0
